KUDU-1236. Maintenance manager schedules empty flushes when server is above memory limit
This patch indirectly addresses this issue by making sure that we
don't consider empty MRS as being flushable. If the MRS is empty,
we now simply skip it when building the list of potential ops to
run. The next time someone is over the soft memory limit and
the MM can't do anything, we'll get a message at INFO level.

While there, I simplified the code a bit which should also save
us a few cycles.

Tested on a local node with YCSB.

Change-Id: I4090986b3d957f767d7346035a69715561ce761f
Reviewed-on: http://gerrit.cloudera.org:8080/1358
Tested-by: Internal Jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>