[java client] Extract the mini cluster from BaseKuduTest
BaseKuduTest does two things: it manages a mini Kudu cluster and it
offers some common utilities to unit tests.

Extracting the mini cluster out of it makes it easier to reuse. This
is particularly helpful for the incoming Spark module since Scala
and static methods don't play nicely together, which BaseKuduTest is
full of (like @BeforeClass).

Change-Id: Ic1b05168d77892e387e280d3569e87586b6ec395
Reviewed-on: http://gerrit.cloudera.org:8080/1365
Tested-by: Internal Jenkins
Reviewed-by: Dan Burkert <dan@danburkert.com>