KUDU-728. tablet_random_access-test should trigger delta compactions, fix bug
This enables minor and major delta compactions in tablet_random-access-test,
and fixes a bug that this exposed.

The bug was that minor delta compaction was using std::sort to sort the deltas
being output to the file, but this didn't respect the ordering of deltas within
a batch if they shared the same timestamp.

Change-Id: I597a6f8763ab8009e4f15245eada9c862d0eef77
Reviewed-on: http://gerrit.cloudera.org:8080/1402
Tested-by: Internal Jenkins
Reviewed-by: David Ribeiro Alves <david.alves@cloudera.com>