Consolidate clock synchronization checks in HybridClock::WalltimeWithError()
This removes the CheckClockSynchronized() method and instead makes
WalltimeWithError() also check that the error is not beyond the
configured bound, consolidating this check.

HybridClock initialization now calls WalltimeWithError() and returns
an error immediately if the clock is not synchronized.

Finally this moves clock initialization in ServerBase to be done
sooner, where CheckClockSynchronized() was previously called.

Change-Id: I8c42d4380d7e7b7ea9e377bc48eaa7b66c7e3a6b
Reviewed-on: http://gerrit.cloudera.org:8080/1420
Reviewed-by: Dan Burkert <dan@danburkert.com>
Reviewed-by: Jean-Daniel Cryans
Tested-by: Internal Jenkins