KUDU-430 - Mock system clock reads in HybridClock
This introduces a new flag, 'use_mock_wall_clock' that, when set,
makes HybridClock not read the actual system clock value and error
and instead read pre-set values.

Note that this does not mock the behavior of HybridClock itself, it
just mocks the behavior of the system clock from which the HybridClock
reads.

This is essential for testing synchronization issues and in general to
be able to emulate clock disparities in tablet servers running in the
same machine.

Change-Id: Ie750b0426bfeb7e8c2a21a6e695cef86a5b59b45
Reviewed-on: http://gerrit.cloudera.org:8080/1421
Tested-by: Internal Jenkins
Reviewed-by: Jean-Daniel Cryans