Add a RPC to set the server's wall clock and error, for tests
This adds a new RPC to GenericService to allow to set the server's
wall clock and max clock error, so that we can do it from integration
tests.

The RPC requires that the server was started with the HybridClock and
the flag '--use_mock_wall_clock= true', otherwise it will respond with
an error (i.e. it doesn't crash the server).

Change-Id: I09c159daffb858d29ef91df298cd679ef493e479
Reviewed-on: http://gerrit.cloudera.org:8080/1422
Tested-by: Internal Jenkins
Reviewed-by: Jean-Daniel Cryans