[cpp/java client] - Clarify the meaning of the read modes
This clarifies the meaning of scanner's read modes in both clients.
We were previously pointing users to the explanation in common.proto,
which was not a good idea, since users may not have direct access to
it.

This also makes it clear what the read modes mean in ACID terms, since
we've been asked that question a lot.

Change-Id: I7512a9c62c7f746b883a670ab48d79fa83febf09
Reviewed-on: http://gerrit.cloudera.org:8080/1427
Reviewed-by: Jean-Daniel Cryans
Tested-by: Internal Jenkins