[cpp client] - Add external consistency mode to the cpp client
This adds ExternalConsistencyMode to the cpp client. It does not
add any tests as there are already tests for the actual consistency semantics
in TabletServerTest. However, now that we have access to the full consistency api
from the client we should move those tests to either client-test
or a new integration test, created KUDU-1263 to track this.

Change-Id: I6f203c829dfa413823172f9884c7cbfd40254f35
Reviewed-on: http://gerrit.cloudera.org:8080/1451
Reviewed-by: Jean-Daniel Cryans
Tested-by: Internal Jenkins