kudu-ts-cli should pass UUID when calling DeleteTablet()
This change adds a unit test for the kudu-ts-cli delete_tablet
functionality, and ensures that the UUID of the tablet server is passed
with the RPC request. This is done by first fetching the UUID from the
remote and then including it in the subsequent request. This allows us
to leave in a DEBUG-only check to ensure that UUID is passed with all
requests, which in most cases makes sense -- even though in this case it
seems like overkill, with a silly workaround.

There is no change to the command-line interface.

Also added another variant of Subprocess::Call().

Change-Id: I9ad1e2f1fe0bbe80e5ad850a4bae45ff110b59db
Reviewed-on: http://gerrit.cloudera.org:8080/1458
Tested-by: Internal Jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>