KUDU-1021 and KUDU-1155. Clarify Status on /tablets page
This patch makes a couple of changes to help clarify the state a replica
is in, as displayed on the /tablets page of the tablet server web app.

Previously, we had some nonintuitive output when a tablet was deleted:
1. A tombstoned tablet would display as SHUTDOWN immediately after
   tombstoning the tablet.
2. A tombstoned tablet would display as BOOTSTRAPPING after a restart
3. Last Status is not updated to reflect that a tablet was deleted

This patch fixes these issues by:
- Differentiating between NOT_STARTED and BOOTSTRAPPING (added another
  runtime state).
- Make the UI reflect not only the TabletStatePB enum values but also
  the TabletDataState enum values as well.

Tested this manually in my dev environment.

Change-Id: I14b878985b4eab520de184c25927f30f5beb7ef5
Reviewed-on: http://gerrit.cloudera.org:8080/1498
Tested-by: Internal Jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>