KUDU-1265. Add a script to check API compatibility between versions
This adds a Python script which downloads and runs the Java API Compatibility
Checker between two arbitrary git revisions.

It's not yet wired into any part of the build, but we can run it manually
prior to a release. I've done so for 0.5.0 vs 0.6.0-rc1 and posted the results
to https://issues.cloudera.org/browse/KUDU-1228

Change-Id: I39d870502b2c8acf9008a122459c4f3ab6dd06a0
Reviewed-on: http://gerrit.cloudera.org:8080/1489
Tested-by: Internal Jenkins
Reviewed-by: Jean-Daniel Cryans