rpc: add a flag for keepalive time, bump default to 65s
We're seeing KUDU-279 fairly often in stress tests on a cluster.
Until that is fixed, it makes sense to bump the RPC keepalive time
significantly. Bumping to 65 seconds should be enough that any
"once a minute" type heartbeats would keep the connection alive.
In particular, this will be long enough that Impala's scanner keepalive
heartbeats will keep RPC connections to the tablet servers alive,
hopefully preventing the errors we're seeing in some large Impala
queries.

Change-Id: I3059ca764447b8073b69972728ad1ebd41b3e8c4
Reviewed-on: http://gerrit.cloudera.org:8080/1497
Reviewed-by: Jean-Daniel Cryans
Tested-by: Internal Jenkins