Convert FAQ to markdown
This page is simple enough to render in markdown without any noticeable
lack of quality and allows us to avoid maintaining the list of links at
the top. I would have used asciidoc instead of markdown, but
unfortunately GitHub doesn't enable that plugin. We could add support
for the asciidoc plugin down the road when we are hosting our own static
site and generating the site automatically with Jenkins.

One downside to auto-generating the TOC is that our anchor links will
change, and direct links to previous entries will break. But the user
will still be able to look for their answer on the FAQ page (they won't
get a 404 Not Found error) so I think it's worth it.

Additionally, add a single_col Jekyll layout for use by markdown-only
pages, and factor the contents of default.html into "top" and "bottom"
includes so that single_col.html can share all of the "chrome" code.
single_col.html is good for simple pages that only need a single column
of basic content.

Change-Id: I453973211eef948d7ab56dcb416f3e39fd8f69da
Reviewed-on: http://gerrit.cloudera.org:8080/1518
Reviewed-by: Dan Burkert <dan@danburkert.com>
Tested-by: Mike Percy <mpercy@cloudera.com>