Add link to Kudu 0.5.0 release
This commit adds a Download link to the web site navigation bar, along
with markdown files for Kudu releases in general and the Kudu 0.5.0
release. The download / release pages are pretty bare-bones at the
moment.

Unfortunately, when adding an additional item to the navigation bar, we
also overflowed the maximum available width and messed up the layout
(the logo was popping up above the navigation bar). So, I added some CSS
hacks to fix this temporarily. Longer term, we need to redesign the
menu. One of the hacks involved hiding the FAQ link if the browser is in
a certain page width range (it's a pretty narrow range).

Change-Id: If76186f8477906344b9ba97f2f8bccffa53add2d
Reviewed-on: http://gerrit.cloudera.org:8080/1528
Reviewed-by: Misty Stanley-Jones <misty@apache.org>
Tested-by: Mike Percy <mpercy@cloudera.com>