Add script to fully build all of the docs for the site
This mostly factors logic out of a Jenkins job (kudu-docs) and into a
script that we can run manually if needed.

This also pulls some ruby-specific logic from the Jenkins job into the
make_docs.sh script, which is where it is actually required, along with
a Gemfile to help manage the dependencies.

Also add a Gemfile.lock so the docs generation script doesn't need to
check for package updates on every run.

Change-Id: I9dcfcf2a14660a3832f5752a1517f96dd2519288
Reviewed-on: http://gerrit.cloudera.org:8080/1549
Reviewed-by: Misty Stanley-Jones <misty@apache.org>
Tested-by: Mike Percy <mpercy@cloudera.com>