[java client] - Current batch may get stuck if inflight batch got error
This commit fix a bug in java client which may cause write operations
stuck forever. Before this commit, if a batch flush is scheduled but
there is already an inflight batch in-progress, we add a callback to
retry flushing the batch after inflight batch is successfully completed,
but if inflight batch have an error, default PASSTHROUGH errback will be
called and this batch will never be flushed. This commit add retry
flushing to both callback and errback.

Also more informations is added to tracing log to help debug.

Change-Id: Icadfe78a1d70ac6490d36b0801c6d4fa30939955
Reviewed-on: http://gerrit.cloudera.org:8080/1565
Reviewed-by: Mike Percy <mpercy@cloudera.com>
Tested-by: Mike Percy <mpercy@cloudera.com>