Check in .ycm_extra_conf.py to use Kudu with YouCompleteMe
This is a configuration file used by YouCompleteMe, a Vim extension for
navigation and code completion with C++ and other languages.
See https://github.com/Valloric/YouCompleteMe

To use this, and add your source directory to the
g:ycm_extra_conf_globlist variable in your .vimrc file. For details on
how to install and configure YouCompleteMe, see the above web site.

Change-Id: I7b3c13bb1d89941e137d20995bbe926a10df7a14
Reviewed-on: http://gerrit.cloudera.org:8080/1628
Reviewed-by: Dan Burkert <dan@danburkert.com>
Tested-by: Mike Percy <mpercy@cloudera.com>