Modify the root directory of gcovr to enable Sonar to parse code coverage
When Sonar analyses the sources it assumes that the paths in the coverage report is
directly corresponding to the location of the source code. Replacing the gcovr root,
from /src to . appears to have no other impact than making it possible for Sonar to
infer which file is being reported.

Change-Id: Ifba3d200ad5798620b184549476cd178aaeaecdf
Reviewed-on: http://gerrit.cloudera.org:8080/1634
Reviewed-by: Martin Grund <mgrund@cloudera.com>
Reviewed-by: Todd Lipcon <todd@cloudera.com>
Tested-by: Todd Lipcon <todd@cloudera.com>