KUDU-1317. Spread creation of new tablets more randomly
This switches the assignment of new tablets on the cluster from being a pure
round-robin to instead use the "power-of-two-choices" algorithm. In this
algorithm, each time we need to select a replica location, we sample a random
two nodes in the cluster, and choose whichever of those has fewer tablets
already on it. Per a simple Python script simulation (attached to the JIRA),
this does a much better job of even distribution than pure random selection,
but also avoids completely overloading a reformatted node with all replicas
of a new table.

This patch aims to solve a problem we're seeing with the round-robin
strategy where recovery is not well parallelized. Because of round-robin,
each tablet server only shares replicas with its "adjacent" nodes in the
round robin ring. So, if a server crashes, at most 4 other servers can
participate in recovery.

To check that the new test isn't flaky, I looped it 1000 times:
http://dist-test.cloudera.org/job?job_id=todd.1453857519.32553

Change-Id: I8a27c2ed52b49baeffb309ebecd3d58192eaeec6
Reviewed-on: http://gerrit.cloudera.org:8080/1654
Reviewed-by: Jean-Daniel Cryans
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Kudu Jenkins