build: TSAN runtime no longer needs -pie
As of the latest versions of LLVM/Clang, the TSAN runtime no longer
requires that the executable be built with -pie. In addition, on new
versions of the Linux kernel (>=4.1), the TSAN runtime fails if the
executable _is_ built with -pie [1].

This patch removes -pie and fixes running TSAN-instrumented tests
on my box with Linux 4.2.

[1] https://github.com/google/sanitizers/issues/503

Change-Id: I78083573c548aac689333370983c1d64e8386406
Reviewed-on: http://gerrit.cloudera.org:8080/1660
Reviewed-by: Dan Burkert <dan@danburkert.com>
Tested-by: Internal Jenkins