Do not add thirdparty/bin to PATH in during builds
We sometimes build a custom Python version in thirdparty, which can make
scripts fail when it accidently ends up on the path.

This also adds the proper flags to TSAN to use the external llvm
symbolizer in thirdparty. As a result, we no longer need
asan_symbolizer.py, and it has been removed.

Change-Id: I0fabe02686eca5e1ec3dbf901d95f36d0443dc41
Reviewed-on: http://gerrit.cloudera.org:8080/1678
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Internal Jenkins