cmake: disable new gcc 5 ABI
Code generation is broken when Kudu is built with gcc 5 on a distro where the
new ABI (new std::list and std::string) is on by default, such as my
machine's Ubuntu 15.10. That's because the precompiled code is built with
clang which doesn't (yet) support abi tags, and due to its dependency on glog,
it winds up with an undefined symbol that cannot be resolved at runtime.

Until clang supports abi tags [1], we must force the use of the old ABI in
Kudu and in any library code that interacts with Kudu.

Besides codegen, the new ABI also raised issues with slice comparison in
diskrowset-test and mt-diskrowset-test. These will also need to be addressed
if the new ABI is to be used in the future.

1. https://llvm.org/bugs/show_bug.cgi?id=23529

Change-Id: If5f4e0d3cd3285a54b185f13daed7463cb85505b
Reviewed-on: http://gerrit.cloudera.org:8080/1683
Tested-by: Dan Burkert <dan@danburkert.com>
Reviewed-by: Dan Burkert <dan@danburkert.com>