Fix tombstoning a tombstoned tablet with the CAS option specified
This is a bug fix for an issue found by Binglin where the master
attempted to tombstone a tablet that had already been tombstoned, and
retried forever. This can happen when an RPC requesting a tombstone
operation times out, but was successful.

Improved the CAS delete unit test to cover this case.

Change-Id: I32ab0d4487378c43fdb77034eb0cffe68c287373
Reviewed-on: http://gerrit.cloudera.org:8080/1685
Tested-by: Internal Jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>