KUDU-1278 - Mark peer responsive when remote bootstrap is in progress
This commit fixes a bug in remote bootstrap where, when a peer starts to remote bootstrap
its last_successful_communication_time is not updated, and so if the remote bootstrap lasts
more than peer timeout time, this peer will be evicted and leader will try to remote bootstrap
on a new peer, which will likely timeout again.
Also during remote bootstrap lots of logs are printed, so some log levels are changed from
INFO to VLOG1.

Change-Id: Ic41f2ad31c437f6ab976987262d8b35f9aa9f3ad
Reviewed-on: http://gerrit.cloudera.org:8080/1759
Reviewed-by: Todd Lipcon <todd@apache.org>
Tested-by: Kudu Jenkins