Delete a tablet explicitly when it is replaced
When master send CreateReplica to all replica peers of a tablet, if any
peer timeout, master will replace the old tablet with a new one, but it
will not delete the old replicas on TS. These relicas will be removed if
TS report them to master, but only when:

  1. Quorum config change.
  2. TS start and make a full report.

If none of those happen, these replicas become orphaned replicas.

Also we may need to add periodic full replica report in TS, to prevent
potential inconsistencies like this, Haven't figure out how to identify
and log the inconsistency accurately, so defer this.

Change-Id: Ib58b1efc334a27e825e68f34dbf3f05c7da60606
Reviewed-on: http://gerrit.cloudera.org:8080/1770
Tested-by: Kudu Jenkins
Reviewed-by: Todd Lipcon <todd@apache.org>