KUDU-815. Fix UBSAN error on Schemas with no columns
This fixes a UBSAN error when calling .data() on an empty vector. In theory,
the C++11 standard allows this, but it appears that our libstdcxx
implementation ends up implementing it on top of vector::front(), which returns
a NULL reference in the case of an empty array. This is, of course, undefined
behavior.

Change-Id: Ie79a430d2478320cd18e67a4e328111fe9afad30
Reviewed-on: http://gerrit.cloudera.org:8080/1789
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Todd Lipcon <todd@apache.org>