Spark connectors for Kudu
This is a cleanup of Ted Malaska's "SparkOnKudu":

https://github.com/tmalaska/SparkOnKudu

Even if this patch is very simple and unfinished, the goal is to
provide a platform for others to add more functionality.

It contains wrappers around common Spark operators that make it easy
to use Kudu. It also has a Data Source implementation.

Basic scala unit tests are provided.

Change-Id: Ic187513ef9724d50024f7401d7ecd19d53554245
Reviewed-on: http://gerrit.cloudera.org:8080/1788
Tested-by: Kudu Jenkins
Reviewed-by: Jean-Daniel Cryans