client_samples-test: wait for processes to exit after kill -9
kill -9 doesn't guarantee immediate delivery of the signal to the target
process. In particular, if the process is in an uninterruptible sleep
state inside the kernel, it may be some small amount of time before
it actually exits. During this time, it's possible it could proceed
and create new files on disk, etc. This might cause occasional failures
of the 'rm -rf' call following the kills in this test case -- a behavior
we're seeing occasionally on Jenkins.

Change-Id: I2ba711723b327168be942405b5354ca57c106043
Reviewed-on: http://gerrit.cloudera.org:8080/1795
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Internal Jenkins