Update to new cpplint, fix warnings
This upgrades to the latest google style guide, and fixes the new warnings. Now
that we're on a newer C++, it makes sense to get the latest guide.

I left the C++11 lint rules turned off in the guide, since it was warning about
use of the <mutex> header. We should decide whether we want to allow this
header before enabling C++11 warnings in the lint tool.

In the process of fixing the warnings, I ran into some issues with the
implementation of ErrnoToCString. Specifically, lint wanted me to include
errno.h from errno.cc, but the weird preprocessor feature #defines previously
used in errno.cc caused all sorts of compile errors when the C++ <string>
header ended up included after it. To avoid these issues, I ended up
re-implementing ErrnoToCString roughly based on the implementation in Mesos[1].

[1] https://reviews.apache.org/r/39005/diff/10#index_header

Change-Id: I197188d5929a4c97c8f138ef790d5df4b8b2be74
Reviewed-on: http://gerrit.cloudera.org:8080/1808
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Kudu Jenkins