build: run local tests parallel with distributed tests
This changes the order of the build slightly so that, if distributed
testing is enabled, it asynchronously submits the job to the dist-test
cluster, and then runs the non-distributed tests locally. When the
local tests are finished, it fetches and processes the results of the
dist-test job.

This shaves several minutes off the precommit build.

Change-Id: Iaac391210589a68b43900deb28abe9aff2fe5c76
Reviewed-on: http://gerrit.cloudera.org:8080/1815
Reviewed-by: Todd Lipcon <todd@apache.org>
Tested-by: Todd Lipcon <todd@apache.org>