build: improve build to better indicate what part failed
The Jenkins build script currently sets an EXIT_STATUS variable
to non-zero when any part of the build has an issue. However,
sometimes it can be difficult when looking at a Jenkins log to
figure out which part failed without scanning through the log.

This adds a failure summary at the bottom which indicates which
part failed, and also adds some nicer headers above each major
part of the build.

Change-Id: I63b7d6c71aecaf61ab36f8fc6cedbfe1f63c6bbe
Reviewed-on: http://gerrit.cloudera.org:8080/1816
Tested-by: Kudu Jenkins
Reviewed-by: Jean-Daniel Cryans