Bump squeasel version
The new version of squeasel goes back to using dlopen for OpenSSL, which fixes
an issue we have hit on Centos 6 with incompatible OpenSSL versions:

Between Centos 6.4 and 6.6 the OpenSSL version was bumped from 1.0.0 to 1.0.1e.
Between the 1.0.0 and 1.0.1e versions, Red Hat introduced symbol versioning for
OpenSSL, which results in any application which is compiled against the newer
version being incompatible with the older version at runtime (or installation
time as the rpm tool knows about the symbol version dependency). As part of moving
to C++11, Kudu recently changed from building releases on Centos 6.4 to Centos
6.6 (and consequently from compiling against OpenSSL 1.0.0 to OpenSSL 1.0.1e).
The result is that Kudu RPMs are failing to install on Centos 6.4. Switching to
dynamic opening of OpenSSL in Squeasel breaks the direct dependency chain, and
allows packaged Kudu to once again work on Centos 6.4.

Change-Id: I5a9903703c3c97e3cc70640a82662ed14dc233fa
Reviewed-on: http://gerrit.cloudera.org:8080/1849
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Internal Jenkins