KUDU-1290. Fix CHECK failures when some CPUs are disabled
This adds a new function base::MaxCPUID() which uses
/sys/devices/system/cpu/present to determine which CPUs are present on the
system. This is an alternative to base::NumCPUs() which doesn't count any CPUs
which have been administratively disabled.

The few places where we use CPU ID to index into a collection are changed
to use this new API.

I verified that, if I disabled one of my CPUs, the master crashed a few minutes
after startup without the patch. With the patch, it seems stable.

Change-Id: I7959e87aba4a694b60ee36adddfd32fef631d8aa
Reviewed-on: http://gerrit.cloudera.org:8080/1864
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Internal Jenkins