client: fix export target munging on Linux
Despite being good intentioned, commit fa6e943 fixed target munging on OS X
at the expense of regressing it on Linux. Specifically, "sed -i ''" yields
backup files with double single-quote suffixes in their names.

After extensive trial and error, the only in-place invocation of sed that'd
work for both platforms required duplicating the call to execute_process().
That's ugly, so here's an approach that foregoes in-place replacement.

Dan and I tested this on Linux and OS X. In both cases:
1. No backup cmake files were created, and
2. The actual replacement took place.

Change-Id: I3451258ff29670367a357111c43e3c27c6392622
Reviewed-on: http://gerrit.cloudera.org:8080/1866
Tested-by: Internal Jenkins
Reviewed-by: Dan Burkert <dan@cloudera.com>