client: Some simplifying cleanups and renames
This patch removes the capability to look up a tablet in the cache by id
via a public API. That API was not actually required by any callers, and
there was a sketchy CHECK in there that enforced a tablet_id hit in the
cache, meaning there were sharp edges that had to be avoided.

Additionally, this patch renames RefreshProxy -> InitProxy and related
functions and removes the "force" bool that was never used by any
callers. There isn't really a reason to "refresh" a proxy unless somehow
the IP address had changed, which we don't currently support. When that
support gets added, the implementer can add whatever APIs make sense at
that time.

Change-Id: I3e825848c6edae69ae5a13e52eaf1e8b9dacb130
Reviewed-on: http://gerrit.cloudera.org:8080/1882
Reviewed-by: Todd Lipcon <todd@apache.org>
Tested-by: Kudu Jenkins