KUDU-1054 (Part 1). Failover client scans on TABLET_NOT_FOUND
If a tablet is deleted from a tablet server, a client that attempts a
scan on the tablet should fail over to another replica if possible.

In addition, this patch allows for marking a tablet as stale in the meta
cache. It replaces RemoteTablet::InvalidateCachedReplicas() with a
MarkStale() method, since InvalidateCachedReplicas() was destructive and
made it difficult to use in a multithreading context.

By adding a "stale" flag to RemoteTablet, we allow users to access
stale replicas until they are replaced (under a lock), which simplifies
the multithreading behavior.

Added a test for failing over client scans that uses tablet deletion and
configuration change to cause a new node to become leader, which
requires updating the tablet locations to scan from.

Also added some test-related helper functions.

Change-Id: Ic97714d82bd943beeb0350babc888b7e5c14ed57
Reviewed-on: http://gerrit.cloudera.org:8080/977
Reviewed-by: Todd Lipcon <todd@apache.org>
Tested-by: Kudu Jenkins