KUDU-1297 - Fix RowSetTree sorting on mac osx
We use a sorting function, RSEndpointBySliceCompare(), for the std::sort()
call in RowSetTree::Reset().

Sorting functions have a hard requirement that they should only return
true iif a < b and should return false in all other cases. Our function would
behave well in all cases except when a == b, in which case it might return
true or false depending on whether a.endpoint_ == START or a.endpoint_ == STOP.

Apparently when this requirement is not met, in some cases, segfaults ensue:
https://schneide.wordpress.com/2010/11/01/bug-hunting-fun-with-stdsort/

The fix is to return false if a == b. I've made sure that this fixes
the failing test (rowset_tree-test) in mac osx.

Change-Id: I315b4a71438cccb7bdc12e8929b064e9e5074f71
Reviewed-on: http://gerrit.cloudera.org:8080/1885
Tested-by: David Ribeiro Alves <david.alves@cloudera.com>
Reviewed-by: Todd Lipcon <todd@apache.org>