client: fix missing includes in scan_batch.h
scan_batch.h was missing some includes for gutil macros that it used.
This didn't cause problems in our own build, because we typically
had included the right files above including this header. But this
broke an Impala build where the caller wasn't explicitly including
other code ahead of this.

This adds a new bit of code to client_samples-test which verifies
that the installed library's headers are all buildable on their own.
This isn't _exactly_ part of the job description of the 'samples' test,
but we already had the build and install code there and didn't want
to duplicate it, nor did I want to take the time to refactor a bunch
of shell code at this juncture.

Change-Id: I8e10f406c80dff7cbc61ca609d8d2bde737b91b6
Reviewed-on: http://gerrit.cloudera.org:8080/1896
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Kudu Jenkins