Use 'kqueue' instead of 'select' for the rpc event loop in Mac OS X
It appears that, in Mac OS X, 'select' can't handle connections when 1024 or
more file descriptors are open. This breaks rpc-test consistently and makes
tablet servers crash when serving a lot of/large tablets in OS X.

This patch makes it so we use 'kqueue' in OS X, which doesn't have the same
problem.

I tested this on my local laptop, rpc-test failed before and passes now.

Change-Id: I8daa3d02c7e28952eecd9e2ce56cadb8b14c418c
Reviewed-on: http://gerrit.cloudera.org:8080/1897
Reviewed-by: Todd Lipcon <todd@apache.org>
Tested-by: Kudu Jenkins