tpch: switch CPU to performance mode before running benchmark
For benchmarks, it's important to disable frequency scaling to get
consistent results. We were already doing this for the main benchmarks
script, but not for the tpch test. One of our benchmark slaves was
set to 'performance' mode whereas another is set to 'ondemand'
which resulted in about a 2x difference in performance on this benchmark.

This refactors out some code for setting the frequency governor and
uses it in the TPCH script.

Change-Id: If917950cef028470d0e603475a236847b1cccd46
Reviewed-on: http://gerrit.cloudera.org:8080/1910
Reviewed-by: Jean-Daniel Cryans
Tested-by: Todd Lipcon <todd@apache.org>