KUDU-1322. Failover client writes on TABLET_NOT_FOUND
Follow-up to KUDU-1054.

If a tablet is deleted from a tablet server, a client that is writing to
the tablet should re-fetch the metadata from the master and connect to
the new leader. If the table has been deleted, the client should get a
Status::NotFound error.

Also grepped for and removed TODOs mentioning KUDU-1054 and KUDU-1322.

Change-Id: I037608977c890cb5099fc74c1be741c93d99ae7a
Reviewed-on: http://gerrit.cloudera.org:8080/978
Tested-by: Kudu Jenkins
Reviewed-by: Jean-Daniel Cryans
Reviewed-by: Todd Lipcon <todd@apache.org>
Reviewed-by: Adar Dembo <adar@cloudera.com>