cmake: blessed build layout fixes and improvements
This patch improves the experience of out-of-tree builds by relying on the
idea of a "blessed" build layout. To summarize, despite now having the
flexibility to place build directories anywhere, we will still encourage
developers to lay them out in a particular way. All scripts and tests will
work well this way, and we'll make only a best effort for them to work
otherwise.

The blessed build layout is where all build directories are subdirectories
of a main build/ subdirectory, itself found in the root directory of the
tree. For example:
- kudu/build/      [unused]
- kudu/build/debug [build directory for debug build]
- kudu/build/asan  [build directory for address sanitizer build]
- kudu/build/tsan  [build directory for thread sanitizer build]

Changes of note:
1. If using the blessed build layout, a build/latest symlink is created (or
   replaced) on every build, pointing to that build's subdirectory.
2. To avoid the confusion of build output in build/latest/latest, binaries
   go in bin/ and libraries in lib/. CMake is smart enough to make the
   corresponding rpath changes to the binaries automatically.
3. Eclipse instructions note that the blessed layout doesn't work properly.
4. Various end-to-end scripts (make_site.sh, tpch.sh, etc.) use the
   appropriate subdirectory in build/, and are more lenient w.r.t. reusing
   existing build output.
5. Java unit tests require no hand-holding for the blessed layout.
6. Python unit tests require just KUDU_HOME for the blessed layout.

Change-Id: I61b4db5a87728d4f3ce9bc13f36d448dbd503c5c
Reviewed-on: http://gerrit.cloudera.org:8080/1914
Reviewed-by: Todd Lipcon <todd@apache.org>
Tested-by: Kudu Jenkins